git-svn-id: http://svn.mkgmap.org.uk/mkgmap/branches/nod@722 25d90789-57f7-4ee0-8453-03a3dfeeeb22

+++ src/uk/me/parabola/imgfmt/app/Section.java
@@ -104,7 +104,7 @@ public class Section {
 		if (itemSize == 0)
 			return 0;
 		
-		return size/itemSize;
+		return size/ (int) itemSize;
 	}
 
 	public void readSectionInfo(ImgFileReader reader, boolean withItemSize) {
